feat[#44]: 이동 거리, 이동 시간 타입 변경
- 이동 거리 : Long -> Double
- 이동 시간 : Long -> Integer

+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hubroute/application/dto/response/HubRouteCreateResponseDto.java
@@ -17,8 +17,8 @@ public class HubRouteCreateResponseDto {
   private UUID endHubId;
   private String startHubName;
   private String endHubName;
-  private Long deliveryTime;
-  private Long deliveryDistance;
+  private Integer deliveryTime;
+  private Double deliveryDistance;
 
   public HubRouteCreateResponseDto(HubRoute hubRoute) {
     this.startHubId = hubRoute.getStartHubId();
