+++ com.sparta.logistics.hub-service/src/main/java/com/sparta/logistics/hub_service/hubroute/application/dto/response/HubRouteListResponseDto.java
@@ -10,8 +10,8 @@ public class HubRouteListResponseDto {
 
   private String startHubName;
   private String endHubName;
-  private Long deliveryTime;
-  private Long deliveryDistance;
+  private Integer deliveryTime;
+  private Double deliveryDistance;
 
   public static HubRouteListResponseDto toResponse(HubRoute hubRoute) {
     return new HubRouteListResponseDto(
